Update for Sentry running Oreo and Pie releases.
Due to changes in Android security and project Treble the pcscd socket
interface that applications use through the libpcsclite.so native
library have moved for Oreo and Pie releases.

This means that your application must be built with a particular pcsc
native library that corresponds to the platform. This is controlled via
the modules build selection which allows you to include a particular
library in your application that corresponds to the Android version as
follows:

For Sentry running:

Lollipop: use module sentrypcsc-debug
Oreo(8): use module sentrypcsc-2.0
Pie(9): use module sentrypcscs-3.0